# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Enterprise HR API Documentation",
        "description": "API Documentation for Team 11 Enterprise HR System, covering Employee, Attendance, and Approval modules.",
        "contact": {
            "email": "admin@enterprise.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Localhost"
        }
    ],
    "paths": {
...
        "/api/employees/{id}/subordinates": {
            "get": {
                "tags": [
                    "Employee"
                ],
                "summary": "Get employee's direct subordinates",
                "operationId": "635b35462113bf4369ffc3e487d60c90",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "manager": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer"
                                                },
                                                "name": {
                                                    "type": "string"
                                                },
                                                "position": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "subordinates": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Employee"
                                            }
                                        },
                                        "count": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Employee not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/employees/{id}/manager": {
            "put": {
                "tags": [
                    "Hierarchy"
                ],
                "summary": "Assign or change a manager for an employee",
                "operationId": "196771ab55fc33ad86733bf96403a580",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Employee ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "manager_id"
                                ],
                                "properties": {
                                    "manager_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Manager assigned successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Employee"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Failed to assign manager (e.g. Circular Reference detected)"
                    },
                    "422": {
                        "description": "Validation failed"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Hierarchy"
                ],
                "summary": "Remove the manager from an employee",
                "operationId": "701509a19128f81d8c678d3386b27a82",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Employee ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Manager removed successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Failed to remove manager"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/employees/{manager_id}/subordinates": {
            "post": {
                "tags": [
                    "Hierarchy"
                ],
                "summary": "Bulk assign subordinates to a manager",
                "operationId": "ac1d7aec7eff3f29f006acf0ccb52c44",
                "parameters": [
                    {
                        "name": "manager_id",
                        "in": "path",
                        "description": "Manager ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "subordinate_ids"
                                ],
                                "properties": {
                                    "subordinate_ids": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        },
                                        "example": [
                                            3,
                                            4,
                                            5
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Subordinates assigned successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Failed to assign subordinates"
                    },
                    "422": {
                        "description": "Validation failed"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/employees/{manager_id}/subordinates/{subordinate_id}": {
            "delete": {
                "tags": [
                    "Hierarchy"
                ],
                "summary": "Remove a specific subordinate from a manager's team",
                "operationId": "c172c7a4f10d6b4ab976d1919299f526",
                "parameters": [
                    {
                        "name": "manager_id",
                        "in": "path",
                        "description": "Manager ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "subordinate_id",
                        "in": "path",
                        "description": "Subordinate ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Subordinate removed successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Failed to remove subordinate"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Employee": {
                "title": "Employee Model",
                "description": "Employee representation",
                "required": [
                    "id",
                    "name",
                    "email",
                    "position",
                    "department",
                    "leave_balance"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "john@enterprise.com"
                    },
                    "position": {
                        "type": "string",
                        "example": "Software Engineer"
                    },
                    "department": {
                        "type": "string",
                        "example": "IT"
                    },
                    "leave_balance": {
                        "type": "integer",
                        "example": 12
                    },
                    "manager_id": {
                        "type": "integer",
                        "example": null,
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T00:00:00.000000Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T00:00:00.000000Z"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}